delete new user account only if not registered yet

+++ controllers/user_actions.py
@@ -204,7 +204,8 @@ def decline_review(): # no auth required
         review.update_record()
 
         user = db.auth_user[review.reviewer_id]
-        db(db.auth_user.id == review.reviewer_id).delete()
+        if user and user.registration_key: # user (auto-created) did not register yet
+            db(db.auth_user.id == review.reviewer_id).delete()
 
         message = "Declined review={}, declineKey={}, user={}".format(review.id, quickDeclineKey, user.email)
